+++ app/routes/components/Avatar/changelog.codegen.ts
@@ -3,4 +3,6 @@
 
 export const changelog = ''
   + '- 2025-09-23: Initial version.\n'
+  + '- 2025-10-03: Add `Avatar.Img` for image support. Update styles to ensure images\n'
+  + '  fit within the avatar circle.\n'
   ;