+++ app/routes/components/Avatar/index.tsx
@@ -1,15 +1,12 @@
-import type { Route } from "./+types/index";
+import { ComponentTemplate } from "@/components/ComponentTemplate";
+import { Section } from "@/components/Section";
 
-import { ComponentTemplate } from "src/components/ComponentTemplate";
-import { Section } from "src/components/Section";
-
-import { sourceCode } from "./source_code.codegen";
-import { demoSourceCode } from "./Demo_source_code.codegen";
 import { changelog } from "./changelog.codegen";
-
 import { Demo } from "./Demo";
+import { demoSourceCode } from "./Demo_source_code.codegen";
+import { sourceCode } from "./source_code.codegen";
 
-export function meta({}: Route.MetaArgs) {
+export function meta() {
   return [
     { title: "Avatar | Copy UI" },
     { name: "description", content: "The Avatar component from Copy UI." },