+++ src/components/Avatar/index.ts
@@ -1,3 +1,3 @@
-// Avatar from copy-ui @ 2025-09-23.
+// Avatar from copy-ui @ 2025-10-03.
 
 export { Avatar } from "./Avatar";