+++ src/components/Avatar/CHANGELOG.md
@@ -1,1 +1,3 @@
 - 2025-09-23: Initial version.
+- 2025-10-03: Add `Avatar.Img` for image support. Update styles to ensure images
+  fit within the avatar circle.